requirements.txt
pandas
requests
tenacity
python-dotenv